Cleanup storage methods - extract SDK 30 methods to new file Context-storage-sdk30.kt - rename create document methods based on the tree uri they use - extract method to get storage ID to standalone method
